fixed int cast double failed error in partitioned and topic item

+++ lib/api/pulsar/pulsar_partitioned_topic_api.dart
@@ -219,9 +219,9 @@ class PulsarPartitionedTopicApi {
     int partitionNum = 0;
     double msgRateIn = 0;
     double msgRateOut = 0;
-    double msgInCounter = 0;
-    double msgOutCounter = 0;
-    double storageSize = 0;
+    int msgInCounter = 0;
+    int msgOutCounter = 0;
+    int storageSize = 0;
 
     if (statsMap.containsKey("metadata")) {
       Map metadata = statsMap["metadata"] as Map;
