+++ lib/vm/pulsar/pulsar_partitioned_topic_basic_view_model.dart
@@ -13,9 +13,9 @@ class PulsarPartitionedTopicBasicViewModel extends BaseLoadViewModel {
   String partitionNum = "";
   double msgRateIn = 0;
   double msgRateOut = 0;
-  double msgInCounter = 0;
-  double msgOutCounter = 0;
-  double storageSize = 0;
+  int msgInCounter = 0;
+  int msgOutCounter = 0;
+  int storageSize = 0;
 
   PulsarPartitionedTopicBasicViewModel(this.pulsarInstancePo, this.tenantResp, this.namespaceResp, this.topicResp);
